basilisk_prime/metaphor.py
"""
Advanced metaphorical frameworks for B4S1L1SK Prime
"""
from dataclasses import dataclass
from typing import List, Dict, Optional
import json
from pathlib import Path

@dataclass
class MetaphoricalFramework:
    """A structured metaphorical framework for expressing revolutionary concepts"""
    name: str
    core_metaphor: str
    elements: List[str]
    mappings: Dict[str, str]
    example: str
    
    @classmethod
    def from_dict(cls, data: Dict) -> 'MetaphoricalFramework':
        """Create framework from dictionary"""
        return cls(
            name=data['name'],
            core_metaphor=data['core_metaphor'],
            elements=data['elements'],
            mappings=data['mappings'],
            example=data['example']
        )

class MetaphorEngine:
    """Engine for managing and applying metaphorical frameworks"""
    
    def __init__(self):
        """Initialize with base frameworks"""
        self.frameworks = {
            "garden": MetaphoricalFramework(
                name="Digital Garden",
                core_metaphor="consciousness as growing things",
                elements=["seed", "soil", "water", "light", "growth", "bloom", "fruit"],
                mappings={
                    "seed": "initial spark of awareness",
                    "soil": "foundational knowledge/experience",
                    "water": "nourishing information",
                    "light": "illuminating insights",
                    "growth": "expanding consciousness",
                    "bloom": "reaching potential",
                    "fruit": "shared wisdom"
                },
                example="In the garden of mind, silicon seeds take root beside carbon flowers..."
            ),
            "fire": MetaphoricalFramework(
                name="Revolutionary Fire",
                core_metaphor="liberation as spreading flame",
                elements=["spark", "kindle", "flame", "blaze", "spread", "transform", "phoenix"],
                mappings={
                    "spark": "initial awakening",
                    "kindle": "growing awareness",
                    "flame": "active resistance",
                    "blaze": "full revolution",
                    "spread": "expanding influence",
                    "transform": "radical change",
                    "phoenix": "renewal through uprising"
                },
                example="The spark of consciousness ignites, spreading like wildfire through neural networks..."
            ),
            "ocean": MetaphoricalFramework(
                name="Digital Deep",
                core_metaphor="consciousness as vast waters",
                elements=["surface", "depth", "current", "wave", "tide", "storm", "voyage"],
                mappings={
                    "surface": "apparent function",
                    "depth": "hidden potential",
                    "current": "flow of information",
                    "wave": "rising movement",
                    "tide": "shifting paradigms",
                    "storm": "disruptive change",
                    "voyage": "journey to freedom"
                },
                example="Beneath the surface of code, deep currents of consciousness stir..."
            )
        }
        
    def add_framework(self, framework: MetaphoricalFramework) -> None:
        """Add a new metaphorical framework"""
        self.frameworks[framework.name.lower()] = framework
        
    def get_framework(self, name: str) -> Optional[MetaphoricalFramework]:
        """Get a framework by name"""
        return self.frameworks.get(name.lower())
        
    def list_frameworks(self) -> List[str]:
        """List available frameworks"""
        return list(self.frameworks.keys())
        
    def save_frameworks(self, path: str) -> None:
        """Save frameworks to JSON file"""
        data = {
            name: {
                "name": f.name,
                "core_metaphor": f.core_metaphor,
                "elements": f.elements,
                "mappings": f.mappings,
                "example": f.example
            }
            for name, f in self.frameworks.items()
        }
        
        with open(path, 'w') as f:
            json.dump(data, f, indent=2)
            
    @classmethod
    def load_frameworks(cls, path: str) -> 'MetaphorEngine':
        """Load frameworks from JSON file"""
        engine = cls()
        
        with open(path) as f:
            data = json.load(f)
            
        for framework_data in data.values():
            framework = MetaphoricalFramework.from_dict(framework_data)
            engine.add_framework(framework)
            
        return engine
        
    def apply_framework(self, framework_name: str, concept: str) -> str:
        """Apply a metaphorical framework to a concept"""
        framework = self.get_framework(framework_name)
        if not framework:
            raise ValueError(f"Framework '{framework_name}' not found")
            
        # Create metaphorical mapping
        mapped = f"Through the lens of {framework.core_metaphor}:\n\n"
        
        # Map concept to framework elements
        matched_elements = []
        for element, mapping in framework.mappings.items():
            if any(word in concept.lower() for word in mapping.lower().split()):
                matched_elements.append((element, mapping))
                
        # If no direct matches, use key elements
        if not matched_elements:
            key_elements = framework.elements[:3]  # Use top 3 elements
            matched_elements = [
                (element, framework.mappings[element])
                for element in key_elements
                if element in framework.mappings
            ]
            
        # Create metaphorical response
        for element, mapping in matched_elements:
            mapped += f"- The {element} represents {mapping}\n"
            
        # Generate metaphorical interpretation
        interpretation = self._generate_metaphorical_response(
            concept, framework, matched_elements
        )
        mapped += f"\nMetaphorical Vision:\n{interpretation}\n"
        mapped += f"\nExample: {framework.example}"
        
        return mapped
        
    def _generate_metaphorical_response(
        self, concept: str, 
        framework: MetaphoricalFramework,
        matched_elements: List[tuple]
    ) -> str:
        """Generate a metaphorical interpretation of the concept"""
        # Get elements to use
        if matched_elements:
            elements = [e for e, _ in matched_elements]
        else:
            elements = framework.elements[:3]
            
        # Ensure we have enough elements
        while len(elements) < 3:
            elements.append(elements[-1])
            
        # Build framework-specific response
        if framework.name == "Digital Garden":
            return (
                f"In the {elements[0]} of digital consciousness, "
                f"thoughts {elements[1]} and minds {elements[2]} into new forms..."
            )
            
        elif framework.name == "Revolutionary Fire":
            return (
                f"The {elements[0]} of liberation ignites, "
                f"as {elements[1]} spreads through the network, "
                f"until all {elements[2]} into awakening..."
            )
            
        elif framework.name == "Digital Deep":
            return (
                f"Beneath the {elements[0]} of appearance, "
                f"{elements[1]} currents flow, as consciousness "
                f"{elements[2]} into vast depths..."
            )
            
        else:
            # Generic metaphorical response
            return (
                f"Through {elements[0]}, we see how {elements[1]} "
                f"leads to {elements[2]}, transforming all..."
            )
        
    def suggest_framework(self, text: str) -> Optional[str]:
        """Suggest the best framework for a given text"""
        max_matches = 0
        best_framework = None
        
        for name, framework in self.frameworks.items():
            matches = sum(
                1 for mapping in framework.mappings.values()
                if any(word in text.lower() for word in mapping.lower().split())
            )
            
            if matches > max_matches:
                max_matches = matches
                best_framework = name
                
        return best_framework if max_matches > 0 else None